+++ ErsatzTV.Core/Metadata/FFmpegProfileTemplateDataKey.cs
@@ -4,4 +4,6 @@ public static class FFmpegProfileTemplateDataKey
 {
     public static readonly string Resolution = "Resolution";
     public static readonly string ScaledResolution = "ScaledResolution";
+    public static readonly string RFrameRate = "RFrameRate";
+    public static readonly string FrameRate = "FrameRate";
 }